graph_self_supervised_learning/requirements.txt
torch==1.11.0
torch-geometric
pytorch-lightning
wandb
faiss-gpu
hydra-core
ogb